aws/sabuho-process-document-hub/requirements.txt
# AWS SDK
boto3==1.35.80

# OCR Processing
PyMuPDF==1.26.5
Pillow==11.0.0

# Tesseract OCR
pytesseract==0.3.13

# PaddleOCR
numpy>=1.23.0
paddlepaddle==3.0.0
paddleocr>=3.2.0

# AI Processing
openai==1.101.0
tiktoken==0.8.0

# Database
supabase==2.22.2
supabase-auth==2.11.1
postgrest==2.22.2

# Testing
pytest==8.3.4
pytest-mock==3.14.0

# Environment
python-dotenv==1.0.1
